Monkey-patch inifile gem not to typecast
With thanks to @brodiepeers for pointing this bug out.

The inifile contains serial numbers, which are typecast to integers by
the inifile gem. If those serial numbers start with a 0 then they are
silently dropped.

This awful monkey-patch replaces the typecast method in inifile with one
that skips the integer/float conversions, so everything is now left as a
string. This does mean things like port numbers are now also strings but
that should be harmless...